refactor(atoms): Moved small components to their own file
